+++ UCR.Core/Models/Subscription/MappingSubscription.cs
@@ -10,20 +10,23 @@ namespace HidWizards.UCR.Core.Models.Subscription
         public List<PluginSubscription> PluginSubscriptions { get; }
         public bool Overriden { get; set; }
 
-        public MappingSubscription(Profile profile, Mapping mapping, Guid subscriptionStateGuid, List<DeviceSubscription> profileOutputDevices)
+        public MappingSubscription(Profile profile, Mapping mapping, Guid subscriptionStateGuid, List<DeviceConfigurationSubscription> subscriptionOutputDeviceConfigurations)
         {
             Mapping = mapping;
             Overriden = false;
             DeviceBindingSubscriptions = new List<InputSubscription>();
             foreach (var mappingDeviceBinding in Mapping.DeviceBindings)
             {
-                DeviceBindingSubscriptions.Add(new InputSubscription(mappingDeviceBinding, profile, subscriptionStateGuid));
+                if (!mappingDeviceBinding.IsBound) continue;
+
+                var inputSubscription = new InputSubscription(mapping, mappingDeviceBinding, profile, subscriptionStateGuid);
+                if (inputSubscription.DeviceSubscription != null) DeviceBindingSubscriptions.Add(inputSubscription);
             }
 
             PluginSubscriptions = new List<PluginSubscription>();
             foreach (var mappingPlugin in Mapping.Plugins)
             {
-                PluginSubscriptions.Add(new PluginSubscription(mappingPlugin, subscriptionStateGuid, profileOutputDevices));
+                PluginSubscriptions.Add(new PluginSubscription(Mapping, mappingPlugin, subscriptionStateGuid, subscriptionOutputDeviceConfigurations));
             }
         }
     }