+++ UCR.Core/Utilities/Constants.cs
@@ -4,7 +4,6 @@ namespace HidWizards.UCR.Core.Utilities
 {
     public static class Constants
     {
-        public const int MaxDevices = 16;
         public const int AxisMaxValue = short.MaxValue;
         public const int AxisMaxAbsValue = short.MinValue * -1;
         public const int AxisMinValue = short.MinValue;