+++ UCR/App.xaml.cs
@@ -141,6 +141,8 @@ namespace HidWizards.UCR
 
         private void App_OnExit(object sender, ExitEventArgs e)
         {
+            context?.DevicesManager.UpdateDeviceCache();
+
             Dispose();
         }